requirements.txt
playwright>=1.40.0
python-telegram-bot>=20.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
loguru>=0.7.2
httpx>=0.25.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
